[feat]: dynamic dao

+++ internal/handler/v1/jwt/types.go
@@ -6,7 +6,7 @@ type JwtHandler interface {
 	//清除token
 	ClearToken(ctx *gin.Context) error
 	ExtractToken(ctx *gin.Context) string
-	SetLoginToken(ctx *gin.Context, uid uint) error
-	SetJwtToken(ctx *gin.Context, uid uint, ssid string) error
+	SetLoginToken(ctx *gin.Context, uid int64) error
+	SetJwtToken(ctx *gin.Context, uid int64, ssid string) error
 	CheckSession(ctx *gin.Context, ssid string) error
 }
